Added new class tree Transformation.
These are designed to have a common base class so that they can be grouped in a container. Each class has a number of member fields, and can produce a Matrix4d on demand, but is not itself a matrix. This makes it easier to modify the transformations on the fly.