Documentapalooza!
Doxygen now no longer reports any undocumented parts. That doesn't mean the documentation is perfect.... Some parts got essay-length essays on parts of ray-tracing, as this program attempts to become a runnable textbook on the subject.

Only one functionality change -- the camera is now more right handed in the most natural way I could come up with. It involved renaming the up vector to be the down vector.